client(feat): related fields chart
